+++ test/recording_studio_support_test.rb
@@ -113,6 +113,8 @@ class RecordingStudioSupportTest < Minitest::Test
     assert_includes tailwind_source, "flatpack-*/app/components/**/*.{rb,erb}"
     assert_includes tailwind_source, "../../../vendor/bundle/**/recording_studio/app/views/**/*.erb"
     assert_includes tailwind_source, "recordingstudio-*/app/views/**/*.erb"
+    assert_includes tailwind_source, "../../../vendor/flat_pack/app/components/**/*.rb"
+    assert_includes tailwind_source, "home/*/.local/share/mise/installs/ruby"
     refute_includes tailwind_source, "@theme"
     refute_includes tailwind_source, ":root {"
     refute_includes tailwind_source, "--color-fp-primary"